add bert teacher progress

+++ src/training/train_bert_teacher.py
@@ -325,12 +325,18 @@ def main(args):
     combined_dataset = Dataset(train=combined_training, test=dataset.test)
 
     if args.save_datasets:
+        logger.info("Saving datasets.")
         data_dir = os.path.join(args.output_dir, 'data')
-        dataset.train.to_csv(os.path.join(data_dir, 'train.csv'), sep='\t')
-        dataset.val.to_csv(os.path.join(data_dir, 'val.csv'), sep='\t')
-        dataset.test.to_csv(os.path.join(data_dir, 'test.csv'), sep='\t')
-        combined_dataset.train.to_csv(os.path.join(data_dir, 'combined_train.csv'), sep='\t')
-        combined_dataset.test.to_csv(os.path.join(data_dir, 'combined_test.csv'), sep='\t')
+        os.makedirs(data_dir, exist_ok=True)
+        # if any csv files already exist, raise an error
+        if any([os.path.exists(os.path.join(data_dir, f)) for f in os.listdir(data_dir)]):
+            raise FileExistsError("Dataset files already exist.")
+        else:
+            dataset.train.to_csv(os.path.join(data_dir, 'train.csv'), sep='\t')
+            dataset.val.to_csv(os.path.join(data_dir, 'val.csv'), sep='\t')
+            dataset.test.to_csv(os.path.join(data_dir, 'test.csv'), sep='\t')
+            combined_dataset.train.to_csv(os.path.join(data_dir, 'combined_train.csv'), sep='\t')
+            combined_dataset.test.to_csv(os.path.join(data_dir, 'combined_test.csv'), sep='\t')
 
     if best_dropout_combination is None:
         raise ValueError("No best dropout combination saved.")
